requirements.txt
pyTelegramBotAPI==4.22.0
groq==0.11.0
python-dotenv==1.0.1
httpx==0.27.0